pyproject.toml
[tool.poetry]
name = "nonvex"
version = "0.0.1"
description = "Flask app for DL hyperparameter optimization"
authors = ["Alec Gunny <alec.gunny@gmail.com>"]

[tool.poetry.scripts]
nonvex = "nonvex.__init__:run_cli"

[tool.poetry.dependencies]
python = ">=3.8,<3.10"

# explicitly adding a libclang dependency for now
# due to this transient issue
# https://pythonissues.com/issues/2674747
libclang = {version = "<12.0", optional = true}

# app dependencies
tensorflow = {version = "^2.6", optional = true}
keras-tuner = {version = "^1.1", optional = true}
flask = {version = "^2.0", optional = true}

# client dependencies
requests = {version = "^2.26", optional = true}
"hermes.typeo" = {version = "^0.1.4", optional = true}

[tool.poetry.extras]
serve = ["tensorflow", "keras-tuner", "flask", "libclang"]
search = ["requests", "hermes.typeo"]

[tool.poetry.dev-dependencies]
pytest = "^6.2"
pytest-flask = "^1.2"
pre-commit = "^2.16"


[tool.black]
line-length = 79
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.tox
  | _build
  | dist
  | docs
)/
'''

[tool.isort]
known_first_party = ["nonvex"]
multi_line_output = 3
include_trailing_comma = true

[build-system]
requires = ["poetry>=1.1"]
build-backend = "poetry.masonry.api"
